Fixed test bug and renamed App Controller to Game Controller for consistency
